@@ -27,14 +27,15 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
         <java.version>11</java.version>
-        <spring-boot.version>2.1.3.RELEASE</spring-boot.version>
+        <spring-boot.version>2.7.8</spring-boot.version>
 <!--        <spring-cloud.version>2021.0.3</spring-cloud.version>-->
 <!--        <spring-cloud-alibaba.version>2021.0.1.0</spring-cloud-alibaba.version>-->
         <!--  持久层配置      -->
-        <mybatis-starter.version>2.1.3</mybatis-starter.version>
-        <druid.version>1.1.14</druid.version>
+        <mysql.version>8.0.32</mysql.version>
+        <mybatis-starter.version>2.3.0</mybatis-starter.version>
+        <druid.version>1.2.8</druid.version>
         <mybstis-plus-starter.version>3.1.1</mybstis-plus-starter.version>
-        <pagehelper.version>1.2.12</pagehelper.version>
+        <pagehelper.version>1.4.5</pagehelper.version>
         <!--  工具类配置      -->
         <hutool.version>5.8.10</hutool.version>
         <guava.version>30.1.1-jre</guava.version>
@@ -97,7 +98,7 @@
             <dependency>
                 <groupId>org.springframework.boot</groupId>
                 <artifactId>spring-boot-dependencies</artifactId>
-                <version>2.1.3.RELEASE</version>
+                <version>${spring-boot.version}</version>
                 <type>pom</type>
                 <scope>import</scope>
             </dependency>
@@ -105,6 +106,11 @@
                 <groupId>org.mybatis.spring.boot</groupId>
                 <artifactId>mybatis-spring-boot-starter</artifactId>
                 <version>${mybatis-starter.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>mysql</groupId>
+                <artifactId>mysql-connector-java</artifactId>
+                <version>${mysql.version}</version>
             </dependency>
             <dependency>
                 <groupId>com.baomidou</groupId>
