@@ -1,15 +1,11 @@
 package com.lijiawei.practice.mymall.learning.init;
 
-import org.apache.ibatis.logging.stdout.StdOutImpl;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.test.context.junit4.SpringRunner;
 
-import java.time.*;
-import java.time.temporal.ChronoUnit;
 
 /**
  * @author Li JiaWei
@@ -18,7 +14,6 @@ import java.time.temporal.ChronoUnit;
  * @Date: 2023/3/20 13:42
  * @Version: 1.0
  */
-@RunWith(SpringRunner.class)
 @SpringBootTest
 public class LocalDateTimeTest {
 
